Type : Feat Add
1. 프로그래밍 응용 4차과제
- Figure의 멤버변수인 posX와 posY가 Double로 변경되었음.
- 출력 시 원형, 삼각형, 사각형 순으로 출력되고, 각각에 대해 면적 오름차순으로 출력되는 기능 추가.
- 데이터를 불러오고 저장할 shape.dat 추가.

+++ 프로그래밍 응용 과제/Chapter 4/Chapter 4/Square.cpp
@@ -41,7 +41,7 @@ void Square::print()
 	ioh.printMessage("¸éÀû :" + std::to_string(area));
 	ioh.printMessage("µÑ·¹ :" + std::to_string(perimeter));
 	ioh.printMessage("²ÀÁþÁ¡ 1 :(" + std::to_string(posX) + ", " + std::to_string(posY) + ")");
-	ioh.printMessage("²ÀÁþÁ¡ 2 :(" + std::to_string(posX) + ", " + std::to_string(posY + (int)height) + ")");
-	ioh.printMessage("²ÀÁþÁ¡ 3 :(" + std::to_string(posX + (int)width) + ", " + std::to_string(posY) + ")");
-	ioh.printMessage("²ÀÁþÁ¡ 4 :(" + std::to_string(posX + (int)width) + ", " + std::to_string(posY + (int)height) + ")");
+	ioh.printMessage("²ÀÁþÁ¡ 2 :(" + std::to_string(posX) + ", " + std::to_string(posY + height) + ")");
+	ioh.printMessage("²ÀÁþÁ¡ 3 :(" + std::to_string(posX + width) + ", " + std::to_string(posY) + ")");
+	ioh.printMessage("²ÀÁþÁ¡ 4 :(" + std::to_string(posX + width) + ", " + std::to_string(posY + height) + ")");
 }
